key_manager::key_manager
Struct KeyManager

pub struct KeyManager { /* private fields */ }

This module provides a key manager for managing BitVMX keys and signatures. It includes functionality for generating, importing, and deriving keys, as well as signing messages using ECDSA, Schnorr and Winternitz algorithms. The key manager uses a secure storage mechanism to store the keys.

Implementations

impl KeyManager

pub fn new( network: Network, key_derivation_path: &str, key_derivation_seed: Option<[u8; 32]>, winternitz_seed: Option<[u8; 32]>, keystore: KeyStore, store: Rc<Storage>, ) -> Result<Self, KeyManagerError>

pub fn import_private_key( &self, private_key: &str, ) -> Result<PublicKey, KeyManagerError>

pub fn import_secret_key( &self, secret_key: &str, network: Network, ) -> Result<PublicKey, KeyManagerError>

pub fn import_partial_secret_keys( &self, partial_keys: Vec<String>, network: Network, ) -> Result<PublicKey, KeyManagerError>

pub fn import_partial_private_keys( &self, partial_keys: Vec<String>, network: Network, ) -> Result<PublicKey, KeyManagerError>

pub fn import_rsa_private_key( &self, private_key: &str, index: usize, ) -> Result<String, KeyManagerError>

pub fn generate_keypair<R: Rng + ?Sized>( &self, rng: &mut R, ) -> Result<PublicKey, KeyManagerError>

pub fn generate_master_xpub(&self) -> Result<Xpub, KeyManagerError>

pub fn derive_keypair(&self, index: u32) -> Result<PublicKey, KeyManagerError>

pub fn derive_public_key( &self, master_xpub: Xpub, index: u32, ) -> Result<PublicKey, KeyManagerError>

pub fn derive_winternitz( &self, message_size_in_bytes: usize, key_type: WinternitzType, index: u32, ) -> Result<WinternitzPublicKey, KeyManagerError>

pub fn derive_multiple_winternitz( &self, message_size_in_bytes: usize, key_type: WinternitzType, initial_index: u32, number_of_keys: u32, ) -> Result<Vec<WinternitzPublicKey>, KeyManagerError>

pub fn generate_rsa_keypair<R: RngCore + CryptoRng>( &self, rng: &mut R, index: usize, ) -> Result<String, KeyManagerError>

pub fn sign_ecdsa_message( &self, message: &Message, public_key: &PublicKey, ) -> Result<Signature, KeyManagerError>

pub fn sign_ecdsa_recoverable_message( &self, message: &Message, public_key: &PublicKey, ) -> Result<RecoverableSignature, KeyManagerError>

pub fn sign_ecdsa_messages( &self, messages: Vec<Message>, public_keys: Vec<PublicKey>, ) -> Result<Vec<Signature>, KeyManagerError>

pub fn sign_ecdsa_recoverable_messages( &self, messages: Vec<Message>, public_keys: Vec<PublicKey>, ) -> Result<Vec<RecoverableSignature>, KeyManagerError>

pub fn sign_schnorr_message( &self, message: &Message, public_key: &PublicKey, ) -> Result<Signature, KeyManagerError>

pub fn sign_schnorr_message_with_tap_tweak( &self, message: &Message, public_key: &PublicKey, merkle_root: Option<TapNodeHash>, ) -> Result<(Signature, PublicKey), KeyManagerError>

pub fn sign_schnorr_message_with_tweak( &self, message: &Message, public_key: &PublicKey, tweak: &Scalar, ) -> Result<(Signature, PublicKey), KeyManagerError>

pub fn sign_schnorr_messages( &self, messages: Vec<Message>, public_keys: Vec<PublicKey>, ) -> Result<Vec<Signature>, KeyManagerError>

pub fn sign_winternitz_message( &self, message_bytes: &[u8], key_type: WinternitzType, index: u32, ) -> Result<WinternitzSignature, KeyManagerError>

pub fn export_secret( &self, pubkey: &PublicKey, ) -> Result<PrivateKey, KeyManagerError>

pub fn sign_rsa_message( &self, message: &[u8], index: usize, ) -> Result<Signature, KeyManagerError>

pub fn encrypt_rsa_message( &self, message: &[u8], pub_key: String, ) -> Result<Vec<u8>, KeyManagerError>

pub fn decrypt_rsa_message( &self, encrypted_message: &[u8], index: usize, ) -> Result<Vec<u8>, KeyManagerError>

pub fn get_key_pair_for_too_insecure( &self, aggregated_pubkey: &PublicKey, ) -> Result<(PrivateKey, PublicKey), KeyManagerError>

pub fn sign_partial_message( &self, aggregated_pubkey: &PublicKey, my_public_key: PublicKey, secnonce: SecNonce, aggregated_nonce: AggNonce, tweak: Option<Scalar>, message: Vec<u8>, ) -> Result<PartialSignature, KeyManagerError>

pub fn generate_nonce_seed( &self, index: u32, public_key: PublicKey, ) -> Result<[u8; 32], KeyManagerError>

pub fn new_musig2_session( &self, participant_pubkeys: Vec<PublicKey>, my_pub_key: PublicKey, ) -> Result<PublicKey, KeyManagerError>

pub fn get_my_public_key( &self, aggregated_pubkey: &PublicKey, ) -> Result<PublicKey, KeyManagerError>

pub fn get_pubkey( &self, aggregated_pubkey: &PublicKey, ) -> Result<PublicKey, KeyManagerError>

pub fn aggregate_nonces( &self, aggregated_pubkey: &PublicKey, id: &str, pub_nonces_map: HashMap<PublicKey, Vec<(MessageId, PubNonce)>>, ) -> Result<(), KeyManagerError>

pub fn get_my_pub_nonces( &self, aggregated_pubkey: &PublicKey, id: &str, ) -> Result<Vec<(MessageId, PubNonce)>, KeyManagerError>

pub fn get_my_pub_nonce( &self, aggregated_pubkey: &PublicKey, id: &str, message_id: &str, ) -> Result<PubNonce, KeyManagerError>

pub fn save_partial_signatures_multi( &self, aggregated_pubkey: &PublicKey, id: &str, partial_signatures_mapping: HashMap<PublicKey, Vec<(MessageId, PartialSignature)>>, ) -> Result<(), KeyManagerError>

pub fn save_partial_signatures( &self, aggregated_pubkey: &PublicKey, id: &str, other_public_key: PublicKey, other_partial_signatures: Vec<(MessageId, PartialSignature)>, ) -> Result<Vec<(MessageId, PartialSignature)>, KeyManagerError>

pub fn get_my_partial_signatures( &self, aggregated_pubkey: &PublicKey, id: &str, ) -> Result<Vec<(MessageId, PartialSignature)>, KeyManagerError>

pub fn get_my_partial_signature( &self, aggregated_pubkey: &PublicKey, id: &str, message_id: &str, ) -> Result<PartialSignature, KeyManagerError>

pub fn get_aggregated_signature( &self, aggregated_pubkey: &PublicKey, id: &str, message_id: &str, ) -> Result<Signature, KeyManagerError>

pub fn generate_nonce( &self, message_id: &str, message: Vec<u8>, aggregated_pubkey: &PublicKey, id: &str, tweak: Option<Scalar>, ) -> Result<(), KeyManagerError>