key_manager::musig2::musig
Trait MuSig2SignerApi

pub trait MuSig2SignerApi {
    // Required methods
    fn new_session(
        &self,
        participant_pubkeys: Vec<PublicKey>,
        my_pub_key: PublicKey,
    ) -> Result<PublicKey, Musig2SignerError>;
    fn aggregate_nonces(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
        pub_nonces_map: HashMap<PublicKey, Vec<(MessageId, PubNonce)>>,
    ) -> Result<(), Musig2SignerError>;
    fn get_my_pub_nonces(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
    ) -> Result<Vec<(MessageId, PubNonce)>, Musig2SignerError>;
    fn get_my_pub_nonce(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
        message_id: &str,
    ) -> Result<PubNonce, Musig2SignerError>;
    fn save_partial_signatures(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
        partial_signatures: HashMap<PublicKey, Vec<(MessageId, PartialSignature)>>,
    ) -> Result<(), Musig2SignerError>;
    fn get_data_for_partial_signatures(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
    ) -> Result<PartialSignatureData, Musig2SignerError>;
    fn get_data_for_partial_signature(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
        message_id: &str,
    ) -> Result<(Vec<u8>, SecNonce, Option<Scalar>, AggNonce), Musig2SignerError>;
    fn verify_partial_signatures(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
        pubkey: PublicKey,
        partial_signatures: Vec<(MessageId, PartialSignature)>,
    ) -> Result<bool, Musig2SignerError>;
    fn verify_final_signature(
        &self,
        message_id: &str,
        final_signature: Signature,
        aggregated_pubkey: PublicKey,
        id: &str,
    ) -> Result<bool, Musig2SignerError>;
    fn get_aggregated_signature(
        &self,
        aggregated_pubkey: &PublicKey,
        id: &str,
        message_id: &str,
    ) -> Result<Signature, Musig2SignerError>;
}

Required Methods

fn new_session( &self, participant_pubkeys: Vec<PublicKey>, my_pub_key: PublicKey, ) -> Result<PublicKey, Musig2SignerError>

Initializes a new MuSig2 signing session.

Arguments

participant_pubkeys - Public keys of all signing participants (will be sorted internally)

my_pub_key - Public key of the current participant

Returns

Returns Ok(()) if initialization succeeds, or an error if:

Less than 2 participants

Current participant’s key not included

fn aggregate_nonces( &self, aggregated_pubkey: &PublicKey, id: &str, pub_nonces_map: HashMap<PublicKey, Vec<(MessageId, PubNonce)>>, ) -> Result<(), Musig2SignerError>

Aggregates public nonces from other participants.

Arguments

id - Session identifier

pub_nonces_map - Map of participant public keys to their public nonces with message IDs

Returns

Ok if nonces are successfully aggregated, error if:

Session not found

Invalid public key

Nonce already exists

fn get_my_pub_nonces( &self, aggregated_pubkey: &PublicKey, id: &str, ) -> Result<Vec<(MessageId, PubNonce)>, Musig2SignerError>

Gets this participant’s public nonces for all messages.

Arguments

id - Session identifier

Returns

Vector of message IDs and public nonces, or error if:

Session not found

Nonces not yet generated

fn get_my_pub_nonce( &self, aggregated_pubkey: &PublicKey, id: &str, message_id: &str, ) -> Result<PubNonce, Musig2SignerError>

fn save_partial_signatures( &self, aggregated_pubkey: &PublicKey, id: &str, partial_signatures: HashMap<PublicKey, Vec<(MessageId, PartialSignature)>>, ) -> Result<(), Musig2SignerError>

Aggregates partial signatures from other participants.

Arguments

id - Session identifier

partial_signatures - Map of participant public keys to their partial signatures with message IDs

Returns

Ok if signatures are successfully aggregated, error if:

Session not found

Invalid public key

Invalid partial signature

Signature already exists

fn get_data_for_partial_signatures( &self, aggregated_pubkey: &PublicKey, id: &str, ) -> Result<PartialSignatureData, Musig2SignerError>

Gets this participant’s partial signatures for all messages.

Arguments

id - Session identifier

Returns

Vector of message IDs and partial signatures, or error if:

Session not found

Incomplete participant nonces

fn get_data_for_partial_signature( &self, aggregated_pubkey: &PublicKey, id: &str, message_id: &str, ) -> Result<(Vec<u8>, SecNonce, Option<Scalar>, AggNonce), Musig2SignerError>

fn verify_partial_signatures( &self, aggregated_pubkey: &PublicKey, id: &str, pubkey: PublicKey, partial_signatures: Vec<(MessageId, PartialSignature)>, ) -> Result<bool, Musig2SignerError>

Verifies partial signatures from a participant.

Arguments

id - Session identifier

pubkey - Public key of participant whose signatures to verify

partial_signatures - Vector of message IDs and partial signatures to verify

Returns

True if all signatures are valid, error if:

Session not found

Invalid partial signature

fn verify_final_signature( &self, message_id: &str, final_signature: Signature, aggregated_pubkey: PublicKey, id: &str, ) -> Result<bool, Musig2SignerError>

Verifies a final signature for a specific message.

Arguments

id - Session identifier

message_id - ID of message to verify signature for

final_signature - Final aggregated signature to verify

aggregated_pubkey - Aggregated public key from init_musig2

Returns

True if signature is valid, error if:

Session not found

Invalid signature

fn get_aggregated_signature( &self, aggregated_pubkey: &PublicKey, id: &str, message_id: &str, ) -> Result<Signature, Musig2SignerError>

Gets the final aggregated signature for a specific message.

Arguments

musig_id - Session identifier

message_id - ID of message to get signature for

Returns

Final aggregated signature, or error if:

Session not found

Message not found

Missing partial signatures