key_manager::musig2
Type Alias PartialSignature

pub type PartialSignature = MaybeScalar;

Partial signatures are just scalars in the range [0, n).

See the documentation of [secp::MaybeScalar] for the parsing, serializing, and conversion traits available on this type.

Aliased Type

enum PartialSignature {
    Zero,
    Valid(Scalar),
}

Variants

Zero

Valid(Scalar)