key_manager::winternitz
Trait HashFunction

pub trait HashFunction {
    // Required methods
    fn hash(&self, data: &WinternitzHash) -> WinternitzHash;
    fn hash_size(&self) -> usize;
}

Core hashing functionality for Winternitz One-Time Signatures.

This trait defines the interface for hash functions used in WOTS. The hash function is applied iteratively W times (where W = 2^NBITS - 1) to create and verify one-time signatures.

Security Note

Winternitz signatures are one-time use only. Reusing a key pair compromises security and allows signature forgery.

Required Methods

fn hash(&self, data: &WinternitzHash) -> WinternitzHash

Compute the hash of the given data.

Arguments

data - The input data to hash

Returns

A WinternitzHash containing the hash output

fn hash_size(&self) -> usize

Returns the size of the hash output in bytes.

Returns

32 for SHA256

20 for HASH160